ceph: MDS client
The MDS client is responsible for submitting requests to the MDS
cluster and parsing the response.  We decide which MDS to submit each
request to based on cached information about the current partition of
the directory hierarchy across the cluster.  A stateful session is
opened with each MDS before we submit requests to it.  If a MDS fails
and/or recovers, we resubmit (potentially) affected requests as
needed.