ceph: kill off ceph_vinop() macro
The macro ceph_vinop() takes a Linux inode and expands it into a
(ceph inode number, ceph snapshot id) pair.  It is used in
printf-style formats, together with a format string (%llx.%llx).

This expansion can be confusing, making the code appear to have a
mismatch between the number of format specifiers and the number of
things to be formatted.  There exists a pair of concise functions to
extract the inode and snapshot id: ceph_ino() and ceph_snap().

Replace all uses of ceph_vinop() with calls to those to functions
instead.

Signed-off-by: Alex Elder <elder@dreamhost.com>