ceph: fix some ceph_decode_need() calls
There are spots where ceph_decode_need() or ceph_encode_need() is
called with the wrong sizes, or not called when it should be.  In
some cases calling a "safe" decode operation is sufficient, but
otherwise change the code so the correct and intended overrun
safety checks are being made.

Move the accounting for the data copied into a ceph_pg_pool out
of osdmap_decode() and osdmap_apply_incremental() and into
__decode_pool() where the it actually gets "consumed."

Define ceph_encode_8_safe()--it's the only one missing from an
otherwise pretty complete set, and it was needed in one spot.

While in there, insert some whitespace to tidy up the alignment of
the line-terminating backslashes in some of the macro definitions.
Signed-off-by: Alex Elder <elder@dreamhost.com>