ceph: truncate snap id explicitly in ceph_getattr()
In ceph_getattr(), a ceph inode's snapshot id is returned in the
stat buffer's dev field.  That field is (at the moment anyway)
a 32-bit value, while a snapshot id is 64 bits.  This means that
if the snapshot id happens reach 2^32 or more, the value passed
to the user will be truncated.  This currently happens silently.

Do the truncation explicitly, to make it obvious that it is
occurring, and issue a warning if it ever does occur.

Signed-off-by: Alex Elder <elder@dreamhost.com>