ceph: define ceph_ino_t to represent ceph inode numbers
Define ceph_ino_t to be an opaque type representing a ceph inode
number.  Make use of it everywhere appropriate to more clearly show
which objects are actually ceph inode numbers (and not simply
arbitrary unsigned 64-bit values).  This includes explicitly casting
the reserved values for the root and .ceph directory inodes to this
type.

To further enforce the abstraction, define macros le64_to_ino(), and
ino_to_le64() for converting between on-disk/over-the-wire ceph
inodes (stored as type _le64) and in-core ceph_ino_t.  Define
ceph_encode_ino_safe() and ceph_decode_ino_safe() as well (though
only the encode function is used), for translating between in-core
ceph inodes and their representation in messages.

Signed-off-by: Alex Elder <elder@dreamhost.com>