kernel rbd:: messenger: fix addr_is_blank() to deal with initial empty address
When addr is first initialized or it is really empty, it does not mean ip = 0.0.0.0, but also ss_family may be not defined. It is neither ipv4 nor ipv6. Thus, we may treat those cases correctly and consider it as empty address. So it should return true, not false. The current version kernel rbd will lead to osd printing warnning: "accept peer addr is really 'peer_addr' socket is 'socket_addr ', which is caused by the kernel rbd not learning it address correctly. As addr is initialized as zero memery so that ss_family is neither AF_INET nor AF_INET6, thus it return false to indicate ip is not empty, but it should be empty.