ceph: don't take s_mutex in ceph_check_caps
This seems to be completely unnecessary. Almost all of this function is
done under the inode->i_ceph_lock, aside from the actual sending of the
message. That doesn't seem to be a reason to hold the s_mutex though.
Drop the s_mutex and just rely on the extant locking.

Signed-off-by: Jeff Layton <jlayton@kernel.org>