netfs: Track pages for buffered I/O in netfs_io_request
Add an xarray to the netfs_io_request struct to use as a buffer.  For
buffered reads and writebacks, the intended source/destination pages are
then added to it, supplemented by additional pages to make the buffer up to
a sufficient size to be able to handle an overlarge RPC, decryption and/or
decompression.

The readahead_expand() function will only expand the requested pageset up
to a point where it runs into an already extant page at either end - which
means that the resulting buffer might not be large enough or may be
misaligned for our purposes.

With this, we can make sure we have a useful buffer and we can splice the
extra pages from it into the pagecache if there are holes we can plug.

The buffer could also be useful in the future to perform RMW cycles when
fixing up after disconnected operation or direct I/O with
smaller-than-preferred granularity.

How the buffering is specified and managed in the netfs_io_request struct
is also changed.  Rather than using a collection of flags and trying to
divine the way to do it from those, use an enum with settings that
explicitly say which buffer is used for what, how the data is copied (if at
all) and, in future, if content crypto is to be applied when data is
transferred.

Signed-off-by: David Howells <dhowells@redhat.com>